de: Remove end-of-sequence checks from deserializeSequence
The new design revolves around visitors visiting the *entirety* of
whatever input comes their way. In this case, 'visitSequence' impls
should visit the *entire* sequence, not just the elements.

As such, impls of 'deserializeSequence' no longer need to check for the
end of the sequence, as that's handled by 'visitSequence'. Once
'visitSequence' successfully finishes, 'deserializeSequence' knows that
we've serialized all of the sequence, and can simply return.