de: Refactor Deserializer API
The 'new' constructor has been changed to 'fromReader' (and 'destroy'
has been changed to 'deinit'). This makes it clear that a deserializer's
being made from a reader.

The 'init' constructor now contains 0 allocations, as it should be.

'fromReader' no longer uses 'readAllAlloc' to get a slice for the token
stream constructor. Instead, the deserializer contains an optional array
list which serves as the buffer for the contents of the reader. This
ensures that the deserializer can deallocate the memory read from the
reader only when necessary.

The 'allocator' field has been removed. The only allocators that are
relevant are the one for Getty (which isn't a part of the deserializer)
and the one for the array list (which holds the allocator within
itself).

'fromSlice' has been changed to 'fromSliceAlloc' to indicate that an
allocation has occured during deserialization from a slice.

A new 'fromSlice' function has been added for more ergonomic and
heap-less deserialziation.