ser, de: Move all interface-related info into impls
This change makes it so that implementing types do not carry any
interface-related information. Instead, the appropriate impl blocks
define them.

Although this is a bit more repetitive, I think it makes the distinction
between the implementing type's API and interface-related APIs more
clear. Additionally, impl blocks may define their own Ok and Error types
that are different, in which case this style would be consistent with
how it would look in that scenario.