de: Remove initialization from reader support
For some , the code within Deserializer.deinit was prohibitively slow
(e.g., when deserializing a list a million times, calling into it would
add 20-30ms to the runtime). The only thing in the method is an if
check, so I'm assuming it's that.

Maybe we can re-add the functions later on if people want, but until
then, the library is fine without it.