de: Update nextKeySeed in MapAccess/StructAccess
StructAccess now properly unescapes keys that need unescaping.

The big change here though is the implementation of isKeyAllocated,
which lets StructAccess to specify to the visitor, on a key-by-key
basis, whether or not a deserialized key was allocated (and thus needs
to be freed on the visitor's side in case of an error).