de: Fix integer deserialization
The optimized route was never being taken because we were checking
the wrong type information. We were looking at the visitor's info
instead of the visitor's value's info. This wound up making every
integer deserialization pass u128 or i128 to parseInt, which is
much slower than if you passed in the actual integer type.