de: Remove string constraint for union variants
This allows us to deserialize into unions that have the untagged
attribute set. In those cases, the variant is actually a Content type,
not a string.

This does mean that we need to allocate for each variant, but most
people use tagged unions anyways so that's only one allocation max.