de: Fix parseInt implementation
New compiler changes caused parseInt to raise the most unhelpful compile
error in all of programming. Thankfully, I managed to crawl my way
towards the issue, which was that the return keyword in parseInt wasn't
being applied to the error values in the switch statement. To fix that,
this commit makes each switch prong a return statement.

Also changed in this commit is the signature of parseInt. Since all the
callsites used a base of 10, parseInt now just passes that base itself
instead of taking it as a parameter.