de: Always allocate strings when parsing
Since strings are duped when they are on the stack, there's no reason to
not always allocate them in the deserializer. Not only is it simpler, it
also speeds up performance! I went from 885ms in my basic benchmarking
test to 708ms.

Note that map keys are still alloc_if_needed since keys aren't
duplicated at all. So we shouldn't waste time allocating them if we
don't need to.